refactor credentials/account management
