+++ cli/medperf/commands/auth/logout.py
@@ -1,5 +1,4 @@
 import medperf.config as config
-from medperf.utils import unset_medperf_user_data
 
 
 class Logout:
@@ -8,6 +7,3 @@ class Logout:
         """Revoke the currently active login state."""
 
         config.auth.logout()
-
-        # unset user's medperf server ID
-        unset_medperf_user_data()
