+++ cli/medperf/commands/view.py
@@ -3,7 +3,7 @@ import json
 from typing import Union
 
 from medperf import config
-from medperf.utils import get_medperf_user_data
+from medperf.account_management import get_medperf_user_data
 from medperf.entities.interface import Entity
 from medperf.exceptions import InvalidArgumentError
