+++ cli/medperf/entities/cube.py
@@ -18,6 +18,7 @@ from medperf.exceptions import (
 )
 import medperf.config as config
 from medperf.comms.entity_resources import resources
+from medperf.account_management import get_medperf_user_data
 
 
 class Cube(Entity, Uploadable, MedperfSchema, DeployableSchema):
@@ -112,7 +113,7 @@ class Cube(Entity, Uploadable, MedperfSchema, DeployableSchema):
             callable: A function for retrieving remote entities with the applied prefilters
         """
         comms_fn = config.comms.get_cubes
-        if "owner" in filters and filters["owner"] == config.current_user["id"]:
+        if "owner" in filters and filters["owner"] == get_medperf_user_data()["id"]:
             comms_fn = config.comms.get_user_cubes
 
         return comms_fn
